requirements.txt
requests~=2.31.0
websocket_client
pandas~=2.1.4
pyyaml~=6.0.1
norenrestapipy~=0.0.22
pyotp~=2.9.0
Telethon~=1.33.1
retry~=0.9.2
SQLAlchemy~=2.0.23
tzdata~=2023.4
cacheout~=0.16.0